进一步分离 Core
