+++ lib/main.dart
@@ -4,17 +4,17 @@ import 'package:app_links/app_links.dart';
 import 'package:bitsdojo_window/bitsdojo_window.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:nyalcf_core_ui/utils/theme_control.dart';
 import 'package:tray_manager/tray_manager.dart';
 import 'package:window_manager/window_manager.dart';
 
-import 'package:nyalcf_core/storages/configurations/launcher_configuration_storage.dart';
 import 'package:nyalcf_core/storages/injector.dart';
 import 'package:nyalcf_core/utils/logger.dart';
 import 'package:nyalcf_core/utils/path_provider.dart';
 import 'package:nyalcf_core/utils/universe.dart';
-import 'package:nyalcf_core/utils/task_scheduler.dart';
+import 'package:nyalcf_core_ui/utils/task_scheduler.dart';
 import 'package:nyalcf_core/utils/deep_link_register.dart';
-import 'package:nyalcf_core/utils/deep_link_executor.dart';
+import 'package:nyalcf_core_ui/utils/deep_link_executor.dart';
 import 'package:nyalcf_ui/main_tray.dart';
 import 'package:nyalcf_ui/main_window.dart';
 import 'package:nyalcf_ui/views/auth/login.dart';
@@ -68,7 +68,7 @@ class _AppState extends State<App> with WindowListener, TrayListener {
   /// 根组件
   @override
   Widget build(BuildContext context) {
-    ThemeData themeData = LauncherConfigurationStorage().getTheme();
+    ThemeData themeData = ThemeControl().getTheme();
 
     return GetMaterialApp(
       logWriterCallback: Logger.getxLogWriter,
