+++ nyalcf_cli/bin/nyalcf_cli.dart
@@ -1,7 +1,9 @@
 import 'package:args/args.dart';
 import 'package:nyalcf_cli/arguments.dart';
-
-const String version = '0.0.1';
+import 'package:nyalcf_core/storages/injector.dart';
+import 'package:nyalcf_core/utils/logger.dart';
+import 'package:nyalcf_core/utils/path_provider.dart';
+import 'package:nyalcf_core/utils/universe.dart';
 
 ArgParser buildParser() {
   return Arguments.all;
@@ -12,7 +14,15 @@ void printUsage(ArgParser argParser) {
   print(argParser.usage);
 }
 
-void main(List<String> arguments) {
+void main(List<String> arguments) async {
+  /// 加载环境
+  await Universe.loadUniverse();
+
+  /// 初始化配置文件等
+  await PathProvider.loadSyncPath();
+  await StoragesInjector.init();
+  await Logger.init();
+
   final ArgParser argParser = buildParser();
   try {
     final ArgResults results = argParser.parse(arguments);
@@ -23,8 +33,13 @@ void main(List<String> arguments) {
       printUsage(argParser);
       return;
     }
+
+    if (results.wasParsed('login')) {
+
+    }
+
     if (results.wasParsed('version')) {
-      print('nyalcf_cli version: $version');
+      print('Nya LoCyanFrp! core version: ${Universe.appVersion}');
       return;
     }
     if (results.wasParsed('verbose')) {
