+++ nyalcf_cli/pubspec.yaml
@@ -1,6 +1,7 @@
 name: nyalcf_cli
 description: A sample command-line application with basic argument parsing.
 version: 0.0.1
+publish_to: none
 # repository: https://github.com/my_org/my_repo
 
 environment:
