+++ nyalcf_core/pubspec.yaml
@@ -1,73 +1,23 @@
 name: nyalcf_core
-description: "Nya LoCyanFrp! core."
-version: 0.0.3
-homepage: https://nyalcf.1l1.icu
-publish_to: none
+description: A starting point for Dart libraries or applications.
+version: 1.0.0
+# repository: https://github.com/my_org/my_repo
 
 environment:
-  sdk: '>=3.3.2 <4.0.0'
-  flutter: ">=1.17.0"
+  sdk: ^3.4.3
 
+# Add regular dependencies here.
 dependencies:
-  flutter:
-    sdk: flutter
-
-  nyalcf_inject:
-    path: ../nyalcf_inject
-  nyalcf_ui:
-    path: ../nyalcf_ui
-
   dio: ^5.4.1
-  get: ^4.6.6
   path_provider: ^2.1.3
   shared_preferences: ^2.2.3
   archive: ^3.4.9
   package_info_plus: ^7.0.0
   logger: ^2.1.0
   crypto: ^3.0.3
-  highlight: ^0.7.0
-  url_launcher: ^6.2.6
   win32_registry: ^1.1.3
-#  synchronized: ^3.1.0+1
+  # path: ^1.8.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^4.0.0
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter packages.
-flutter:
-
-  # To add assets to your package, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-  #
-  # For details regarding assets in packages, see
-  # https://flutter.dev/assets-and-images/#from-packages
-  #
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware
-
-  # To add custom fonts to your package, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts in packages, see
-  # https://flutter.dev/custom-fonts/#from-packages
+  lints: ^3.0.0
+  test: ^1.24.0
