+++ nyalcf_core_ui/.metadata
@@ -4,7 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: "68bfaea224880b488c617afe30ab12091ea8fa4e"
+  revision: "761747bfc538b5af34aa0d3fac380f1bc331ec49"
   channel: "stable"
 
-project_type: package
+project_type: module
