+++ nyalcf_core_ui/README.md
@@ -1,0 +1,4 @@
+# Nya LoCyanFrp! Core UI
+
+包含了 Nya LoCyanFrp! 的核心，这些内容只在UI中使用。
+更多请参阅 NyaLCF Core
