+++ nyalcf_core_ui/lib/utils/frpc/startup_loader.dart
@@ -1,9 +1,9 @@
 import 'package:get/get.dart';
 
-import 'package:nyalcf_core/controllers/user_controller.dart';
+import 'package:nyalcf_ui/controllers/user_controller.dart';
 import 'package:nyalcf_core/storages/configurations/autostart_proxies_storage.dart';
 import 'package:nyalcf_core/utils/frpc/path_provider.dart';
-import 'package:nyalcf_core/utils/frpc/process_manager.dart';
+import 'process_manager.dart';
 import 'package:nyalcf_core/utils/logger.dart';
 
 class FrpcStartUpLoader {
