+++ nyalcf_core_ui/lib/utils/proxies_getter.dart
@@ -1,6 +1,6 @@
 import 'package:get/get.dart';
 
-import 'package:nyalcf_core/controllers/proxies_controller.dart';
+import 'package:nyalcf_ui/controllers/proxies_controller.dart';
 import 'package:nyalcf_core/models/user_info_model.dart';
 import 'package:nyalcf_core/storages/prefs/user_info_prefs.dart';
 import 'package:nyalcf_core/storages/stores/proxies_storage.dart';
