+++ nyalcf_ui/lib/controllers/console_controller.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import 'package:nyalcf_core/utils/frpc/process_manager.dart';
+import 'package:nyalcf_core_ui/utils/frpc/process_manager.dart';
 import 'package:nyalcf_ui/views/panel/console.dart';
 
 /// 控制台 GetX 状态控制器
