+++ nyalcf_ui/lib/controllers/panel_controller.dart
