+++ nyalcf_ui/lib/models/frpc_download_tip.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import 'package:nyalcf_core/controllers/frpc_setting_controller.dart';
+import 'package:nyalcf_ui/controllers/frpc_setting_controller.dart';
 import 'package:nyalcf_core/storages/configurations/frpc_configuration_storage.dart';
 import 'package:nyalcf_ui/models/frpc_download_dialog.dart';
 import 'package:nyalcf_core/utils/frpc/path_provider.dart';
