+++ nyalcf_ui/lib/views/home.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import 'package:nyalcf_core/controllers/console_controller.dart';
-import 'package:nyalcf_core/controllers/frpc_controller.dart';
-import 'package:nyalcf_core/controllers/user_controller.dart';
+import 'package:nyalcf_ui/controllers/console_controller.dart';
+import 'package:nyalcf_ui/controllers/frpc_controller.dart';
+import 'package:nyalcf_ui/controllers/user_controller.dart';
 import 'package:nyalcf_core/models/user_info_model.dart';
 import 'package:nyalcf_core/storages/stores/user_info_storage.dart';
-import 'package:nyalcf_core/utils/frpc/startup_loader.dart';
+import 'package:nyalcf_core_ui/utils/frpc/startup_loader.dart';
 import 'package:nyalcf_core/utils/logger.dart';
 import 'package:nyalcf_core/network/dio/auth/auth.dart';
-import 'package:nyalcf_core/utils/proxies_getter.dart';
+import 'package:nyalcf_core_ui/utils/proxies_getter.dart';
 import 'package:nyalcf_inject/nyalcf_inject.dart';
 import 'package:nyalcf_ui/models/appbar_actions.dart';
 import 'package:nyalcf_ui/models/floating_action_button.dart';
