+++ nyalcf_ui/lib/views/panel/proxies.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import 'package:nyalcf_core/utils/widget/after_layout.dart';
-import 'package:nyalcf_core/controllers/proxies_controller.dart';
-import 'package:nyalcf_core/controllers/user_controller.dart';
+import 'package:nyalcf_core_ui/utils/widget/after_layout.dart';
+import 'package:nyalcf_ui/controllers/proxies_controller.dart';
+import 'package:nyalcf_ui/controllers/user_controller.dart';
 import 'package:nyalcf_core/storages/stores/proxies_storage.dart';
 import 'package:nyalcf_inject/nyalcf_inject.dart';
 import 'package:nyalcf_ui/models/account_dialog.dart';
