+++ nyalcf_ui/lib/views/setting/injector.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import 'package:nyalcf_core/controllers/frpc_setting_controller.dart';
-import 'package:nyalcf_core/controllers/launcher_setting_controller.dart';
+import 'package:nyalcf_ui/controllers/frpc_setting_controller.dart';
+import 'package:nyalcf_ui/controllers/launcher_setting_controller.dart';
 import 'package:nyalcf_inject/nyalcf_inject.dart';
 import 'package:nyalcf_ui/models/appbar_actions.dart';
 import 'package:nyalcf_ui/models/floating_action_button.dart';
