+++ nyalcf_ui/lib/views/setting/launcher_setting.dart
@@ -5,14 +5,14 @@ import 'package:get/get.dart';
 import 'package:open_filex/open_filex.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import 'package:nyalcf_core/controllers/launcher_setting_controller.dart';
+import 'package:nyalcf_ui/controllers/launcher_setting_controller.dart';
 import 'package:nyalcf_core/storages/configurations/launcher_configuration_storage.dart';
 import 'package:nyalcf_core/utils/logger.dart';
 import 'package:nyalcf_core/network/dio/launcher/launcher.dart';
 import 'package:nyalcf_core/utils/path_provider.dart';
-import 'package:nyalcf_core/utils/theme_control.dart';
+import 'package:nyalcf_core_ui/utils/theme_control.dart';
 import 'package:nyalcf_core/utils/universe.dart';
-import 'package:nyalcf_core/tasks/updater.dart';
+import 'package:nyalcf_core_ui/tasks/updater.dart';
 import 'package:nyalcf_inject/nyalcf_inject.dart';
 
 class LauncherSetting {
